requirements.txt
opencv-python==4.1.1.26
numpy==1.17.2
Pillow==2.2.1
opencv-contrib-python==4.1.1.26
pyobjc-framework-Quartz==5.2